Fix typed_substate_layout for ComponentMethodRoyalty
